django/pyproject.toml
[tool.poetry]
name = "mapa"
version = "0.1.0"
description = ""
authors = ["Keith Moss <keithamoss@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.1.5"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
django-model-utils = "^4.3.1"
gunicorn = "^20.1.0"
gevent = "^22.10.1"
greenlet = "^1.1.3"
social-auth-app-django = "^5.0.0"
social-auth-core = "^4.3.0"
sentry-sdk = "^1.12.0"
django-simple-history = "^3.2.0"
jsonschema = "^4.17.3"
djangorestframework-gis = "^1.0"
django-filter = "21.1"
djangorestframework-csv = "^2.1.1"
pytz = "^2022.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
